* also moved EULA check before download to fail-fast

+++ minecraft-server/start-minecraft.sh
@@ -1,5 +1,20 @@
 #!/bin/bash
 
+if [ ! -e /data/eula.txt ]; then
+  if [ "$EULA" != "" ]; then
+    echo "# Generated via Docker on $(date)" > eula.txt
+    echo "eula=$EULA" >> eula.txt
+  else
+    echo ""
+    echo "Please accept the Minecraft EULA at"
+    echo "  https://account.mojang.com/documents/minecraft_eula"
+    echo "by adding the following immediately after 'docker run':"
+    echo "  -e EULA=TRUE"
+    echo ""
+    exit 1
+  fi
+fi
+
 case $VERSION in
   LATEST)
     export VERSION=`wget -O - https://s3.amazonaws.com/Minecraft.Download/versions/versions.json | jsawk -n 'out(this.latest.release)'`
@@ -19,43 +34,45 @@ fi
 
 if [ ! -e server.properties ]; then
   cp /tmp/server.properties .
-fi
 
-if [ -n "$MOTD" ]; then
-  sed -i "/motd\s*=/ c motd=$MOTD" /data/server.properties
-fi
+  if [ -n "$MOTD" ]; then
+    sed -i "/motd\s*=/ c motd=$MOTD" /data/server.properties
+  fi
 
-if [ -n "$LEVEL" ]; then
-  sed -i "/level-name\s*=/ c level-name=$LEVEL" /data/server.properties
-fi
+  if [ -n "$LEVEL" ]; then
+    sed -i "/level-name\s*=/ c level-name=$LEVEL" /data/server.properties
+  fi
 
-if [ -n "$SEED" ]; then
-  sed -i "/level-seed\s*=/ c level-seed=$SEED" /data/server.properties
-fi
+  if [ -n "$SEED" ]; then
+    sed -i "/level-seed\s*=/ c level-seed=$SEED" /data/server.properties
+  fi
 
-if [ -n "$MODE" ]; then
-  case ${MODE,,?} in
-    0|1|2|3)
-      ;;
-    s*)
-      MODE=0
-      ;;
-    c*)
-      MODE=1
-      ;;
-    *)
-      echo "ERROR: Invalid game mode: $MODE"
-      exit 1
-      ;;
-  esac
+  if [ -n "$MODE" ]; then
+    case ${MODE,,?} in
+      0|1|2|3)
+        ;;
+      s*)
+        MODE=0
+        ;;
+      c*)
+        MODE=1
+        ;;
+      *)
+        echo "ERROR: Invalid game mode: $MODE"
+        exit 1
+        ;;
+    esac
   
-  sed -i "/gamemode\s*=/ c gamemode=$MODE" /data/server.properties
+    sed -i "/gamemode\s*=/ c gamemode=$MODE" /data/server.properties
+  fi
 fi
 
-if [ -n "$OPS" ]; then
+
+if [ -n "$OPS" -a ! -e ops.txt.converted ]; then
   echo $OPS | awk -v RS=, '{print}' >> ops.txt
 fi
-if [ -n "$ICON" ]; then
+
+if [ -n "$ICON" -a ! -e server-icon.png ]; then
   echo "Using server icon from $ICON..."
   # Not sure what it is yet...call it "img"
   wget -q -O /tmp/icon.img $ICON
@@ -68,19 +85,5 @@ if [ -n "$ICON" ]; then
   fi
 fi
 
-if [ ! -e /data/eula.txt ]; then
-  if [ "$EULA" != "" ]; then
-    echo "# Generated via Docker on $(date)" > eula.txt
-    echo "eula=$EULA" >> eula.txt
-  else
-    echo ""
-    echo "Please accept the Minecraft EULA at"
-    echo "  https://account.mojang.com/documents/minecraft_eula"
-    echo "by adding the following immediately after 'docker run':"
-    echo "  -e EULA=TRUE"
-    echo ""
-    exit 1
-  fi
-fi
-
 java $JVM_OPTS -jar minecraft_server.$VERSION.jar
+
